+++ src/shapes.ts
@@ -115,6 +115,11 @@ export class Text extends Shape {
       this.scale = scale;
   }
 
+  public setBoundingBox(newWidth: number, newHeight: number): void {
+    this.boundingBox.size.x = newWidth ? newWidth : this.boundingBox.size.x;
+    this.boundingBox.size.y = newHeight;
+  }
+
   private getXCornersUnscaled(xFirstCorner: number, xSecondCorner: number, xMinMaxFactor: number): [number, number] {
     if (this.align == "center") return [xFirstCorner * 0.99, xSecondCorner * 1.01];
     if (["right", "end"].includes(this.align)) return [xFirstCorner, xSecondCorner != 0 ? xSecondCorner * (1 - xMinMaxFactor) : -Math.sign(this.scale.x)];
@@ -443,7 +448,7 @@ export class Text extends Shape {
 
 export class Label extends Shape {
   public shapeSize: Vertex = new Vertex(30, C.MAX_LABEL_HEIGHT);
-  public legend: Shape;
+  public legend: LineSegment | Point | Rect;
   public maxWidth: number = 150;
   constructor(
     shape: Shape,
@@ -452,7 +457,6 @@ export class Label extends Shape {
   ) {
     super();
     this.isScaled = false;
-    this.text.width = this.maxWidth - this.shapeSize.x;
     this.getShapeStyle(shape, this.origin);
     this.buildPath();
   }
@@ -464,7 +468,7 @@ export class Label extends Shape {
 
   public getBounds(): [Vertex, Vertex] { return [null, null] }
 
-  protected buildUnscaledPath(context: CanvasRenderingContext2D) {
+  public buildUnscaledPath(context: CanvasRenderingContext2D): Path2D {
     const matrix = context.getTransform();
     context.resetTransform();
     this.buildPath();
@@ -508,7 +512,7 @@ export class Label extends Shape {
     if (this.legend instanceof Rect) this.legend.size.y = height
     else if (this.legend instanceof LineSegment) this.legend.end.y = this.legend.origin.y + height
     else if (this.legend instanceof Point) this.legend.size = height;
-    this.text.fontsize = height;
+    this.text.setBoundingBox(this.maxWidth - this.shapeSize.x - C.LABEL_TEXT_OFFSET, height);
   }
 
   public static deserialize(data: any, scale: Vertex = new Vertex(1, 1)): Label {
@@ -517,6 +521,7 @@ export class Label extends Shape {
     text.isScaled = false;
     text.baseline = "middle";
     text.align = "start";
+    text.multiLine = true;
     return new Label(new Rect(), text)
   }
 
@@ -541,16 +546,18 @@ export class Label extends Shape {
   private drawText(context: CanvasRenderingContext2D): void {
     const contextMatrix = context.getTransform();
     context.resetTransform();
-    this.text.draw(context);
+    this.text.fontsize = null;
+    this.text.rowIndices = [];
+    try { 
+      this.text.draw(context);
+    } catch (TypeError) {}  // This is to avoid a bug when sizes of Label's objects is too small
     context.setTransform(contextMatrix);
   }
 
   protected drawMembers(context: CanvasRenderingContext2D): void { this.drawText(context) }
 
   public isPointInShape(context: CanvasRenderingContext2D, point: Vertex): boolean {
-    return this.legend.isFilled
-      ? context.isPointInPath(this.path, point.x, point.y)
-      : (context.isPointInPath(this.path, point.x, point.y) || context.isPointInStroke(this.path, point.x, point.y));
+    return context.isPointInPath(this.path, point.x, point.y)  // TODO: fix this to include legend shape
   }
 }