+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.26.0]
+### Add
+- 
+
 ### Fix
+- Labels: 
+    - fix wide text
+    - remove shape hovering
 - RubberBand: fix rubberband deletion when putting it outside view point